flattenProducts() function
using our flattenProducts() function and mapping through each item.
We reset the image object to a variable set to the url link directly.

Then we return the spreaded item as first argument and the new image we created as the 2nd argument, overriding the image object and turning it from an image object, into a object key with a value set to the url link. (item.image.url)