Cart button toggle and styling
Created Cart.svelte component that is shown whenever globalStore.js cart object is true.

App.svelte has the if statement to show cart is true or not.

Big and Small navbar both contain the CartButton.svelte component, which that component has the function to .toggleItem('cart', true), so it opens whenever it's clicked.

Cart itself has a close button to make 'cart:false', so it closes on the if logic placed inside App.svelte