Login.svelte update
Created a basic SignIn/Register page.
Takes in Email/Password/Username.

If function if member is already a user or not.

If member already has an account, it will show Email/Password forms only and bottom text asks 'Need To Register?'
Else, it will show Email/Password/Username to register a new account and bottom text asks 'Already a Member?'