product.js fix
Course lesson is old and lecturer messed up. Figuring it out on my own since this change.

Strapi v3 to v4 change creates this issue where the JSON data is not spread out the same way the video lectures show. Objects are now structured a bit more complex.

Currently made it so the product images uploaded onto strapi database now shows correctly.

Still need to figure out how to get title/price/description output.

Main issue is in Featured.svelte and Products.svelte is our Product prop which is a key is passed through but doesn't meet the requirements for the code to follow through the JSON data easily.