setupUser.js, when user logins in/logsout
setupuser.js declares the username and jwt token if logged in. If logged in, then login button turns into logout button.
Localstorage will provide the username/token inside the database.
After logging in, we navigate the page to the products page.

If user logouts, then the local storage is cleared and the username/jwt is set to null.