Alerts.svelte setting up alerts for users
changes in globalStore.js, const globalStore.
Added 2 more items with boolean values and an another item with default string.

const store's toggleItem will now take in 2 more parameters found within the globalStore. alertText and alertDanger set to false.

if the item is equal to 'alert' which is found in Alert.svelte's handlClose's code block, then we update those values within our storeValue.

These parameters change when the on:click's handClose function is executed by clicking the close alert button.

Inside Login.svelte, we are implementing the globalStore.toggleItem function in our if statement for when our user logs in or enters the incorrect password. (Line34)

If user logs in, an item's value (alert) changes to true, alertText is set to 'welcome to shopping madness'

If user enters incorrect password/errors, it will take in the same parameters but the text will be set to 'there was an error, and the alertDanger is set to true, meaning the alert box color is red.