Login.svelte error functionality
If user inputs incorrect password, it will output an alert. But then user user can spam the submit button as it doesn't gray out after the first failed attempt of logging in.

in our reactive isEmpty, we provide it another condition, the $globalStore.alert. If the alert pops up, then the isEmpty is true, which in turn makes the button's disabled and class:disabled to true in its own case.