stripe functionality fix
If logged in user accessed checkout page with 0 items in cart, old code would produce console error that stripe isn't functioning correctly.

So, I put our stripe elements within a if statement to where if cartTotal is less than 0, or has 0 items, the stripe function would know not to try and execute, producing an error code.